Refacgtor out route and socket defs
