+++ App/socket_mcp.py
@@ -1,0 +1,28 @@
+from app import socketio
+
+from flask import current_app, request
+
+
+def init_socket_mcp(app):
+    route_type = "socket"
+    namespace = "/WebMCP"
+    print(f"Initializing {route_type} handling for {namespace}")
+    
+    @socketio.on("checkInRequested", namespace=namespace)
+    def checkInRequested():
+        socketio.emit("checkIn", namespace=namespace)
+
+
+    @socketio.on("connect", namespace=namespace)
+    def watchdog_connect():
+        app.data.console_queue.put("watchdog connected")
+        app.data.console_queue.put(request.sid)
+        socketio.emit("connect", namespace=namespace)
+        if app.mcpthread == None:
+            app.data.console_queue.put("going to start mcp thread")
+            app.mcpthread = socketio.start_background_task(
+                app.data.mcpProcessor.start, current_app._get_current_object()
+            )
+            app.data.console_queue.put("created mcp thread")
+            app.mcpthread.start()
+            app.data.console_queue.put("started mcp thread")
