+++ main.py
@@ -5,7 +5,6 @@ monkey.patch_all()
 
 from app import app, socketio
 
-import json
 import os
 import socket
 import threading
@@ -14,19 +13,11 @@ import webbrowser
 
 import schedule
 
-from flask import (
-    jsonify,
-    render_template,
-    current_app,
-    request,
-    send_file,
-    send_from_directory,
-)
-from flask_mobility.decorators import mobile_template
-from werkzeug.utils import secure_filename
-
 from App.data import init_data
 from App.route import init_route
+from App.socket_maslowcnc import init_socket_maslowcnc
+from App.socket_maslowcnclogs import init_socket_maslowcnclogs
+from App.socket_mcp import init_socket_mcp
 
 from Background.UIProcessor import UIProcessor  # do this after socketio is declared
 from Background.LogStreamer import LogStreamer  # do this after socketio is declared
@@ -73,775 +64,10 @@ app.mcpthread = None
 ## logstreamerthread set to None.. will be activated upon first websocket connection from log streamer browser
 app.logstreamerthread = None
 
-# init_route(app)
-
-@app.route("/")
-@mobile_template("{mobile/}")
-def index(template):
-    app.data.logger.resetIdler()
-    macro1Title = (app.data.config.getValue("Maslow Settings", "macro1_title"))[:6]
-    macro2Title = (app.data.config.getValue("Maslow Settings", "macro2_title"))[:6]
-    if template == "mobile/":
-        return render_template(
-            "frontpage3d_mobile.html",
-            modalStyle="modal-lg",
-            macro1_title=macro1Title,
-            macro2_title=macro2Title,
-        )
-    else:
-        return render_template(
-            "frontpage3d.html",
-            modalStyle="mw-100 w-75",
-            macro1_title=macro1Title,
-            macro2_title=macro2Title,
-        )
-
-
-@app.route("/controls")
-@mobile_template("/controls/{mobile/}")
-def controls(template):
-    app.data.logger.resetIdler()
-    macro1Title = (app.data.config.getValue("Maslow Settings", "macro1_title"))[:6]
-    macro2Title = (app.data.config.getValue("Maslow Settings", "macro2_title"))[:6]
-    if template == "/controls/mobile/":
-        return render_template(
-            "frontpage3d_mobilecontrols.html",
-            modalStyle="modal-lg",
-            isControls=True,
-            macro1_title=macro1Title,
-            macro2_title=macro2Title,
-        )
-    else:
-        return render_template(
-            "frontpage3d.html",
-            modalStyle="mw-100 w-75",
-            macro1_title=macro1Title,
-            macro2_title=macro2Title,
-        )
-
-
-@app.route("/text")
-@mobile_template("/text/{mobile/}")
-def text(template):
-    app.data.logger.resetIdler()
-    macro1Title = (app.data.config.getValue("Maslow Settings", "macro1_title"))[:6]
-    macro2Title = (app.data.config.getValue("Maslow Settings", "macro2_title"))[:6]
-    if template == "/text/mobile":
-        return render_template(
-            "frontpageText_mobile.html",
-            modalStyle="modal-lg",
-            isControls=True,
-            macro1_title=macro1Title,
-            macro2_title=macro2Title,
-        )
-    else:
-        return render_template(
-            "frontpageText.html",
-            modalStyle="mw-100 w-75",
-            macro1_title=macro1Title,
-            macro2_title=macro2Title,
-        )
-
-
-@app.route("/logs")
-@mobile_template("/logs/{mobile/}")
-def logs(template):
-    print("here")
-    app.data.logger.resetIdler()
-    if template == "/logs/mobile/":
-        return render_template("logs.html")
-    else:
-        return render_template("logs.html")
-
-
-@app.route("/maslowSettings", methods=["POST"])
-def maslowSettings():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        result = request.form
-        app.data.config.updateSettings("Maslow Settings", result)
-        message = {"status": 200}
-        resp = jsonify(message)
-        resp.status_code = 200
-        return resp
-
-
-@app.route("/advancedSettings", methods=["POST"])
-def advancedSettings():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        result = request.form
-        app.data.config.updateSettings("Advanced Settings", result)
-        message = {"status": 200}
-        resp = jsonify(message)
-        resp.status_code = 200
-        return resp
-
-
-@app.route("/webControlSettings", methods=["POST"])
-def webControlSettings():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        result = request.form
-        app.data.config.updateSettings("WebControl Settings", result)
-        message = {"status": 200}
-        resp = jsonify(message)
-        resp.status_code = 200
-        return resp
-
-
-@app.route("/cameraSettings", methods=["POST"])
-def cameraSettings():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        result = request.form
-        app.data.config.updateSettings("Camera Settings", result)
-        message = {"status": 200}
-        resp = jsonify(message)
-        resp.status_code = 200
-        return resp
-
-
-@app.route("/gpioSettings", methods=["POST"])
-def gpioSettings():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        result = request.form
-        app.data.config.updateSettings("GPIO Settings", result)
-        message = {"status": 200}
-        resp = jsonify(message)
-        resp.status_code = 200
-        return resp
-
-
-@app.route("/uploadGCode", methods=["POST"])
-def uploadGCode():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        result = request.form
-        directory = result["selectedDirectory"]
-        # print(directory)
-        f = request.files.getlist("file[]")
-        print(f)
-        home = app.data.config.getHome()
-        app.data.config.setValue(
-            "Computed Settings", "lastSelectedDirectory", directory
-        )
-
-        if len(f) > 0:
-            firstFile = f[0]
-            for file in f:
-                app.data.gcodeFile.filename = (
-                    home
-                    + "/.WebControl/gcode/"
-                    + directory
-                    + "/"
-                    + secure_filename(file.filename)
-                )
-                file.save(app.data.gcodeFile.filename)
-            app.data.gcodeFile.filename = (
-                home
-                + "/.WebControl/gcode/"
-                + directory
-                + "/"
-                + secure_filename(firstFile.filename)
-            )
-            returnVal = app.data.gcodeFile.loadUpdateFile()
-            if returnVal:
-                message = {"status": 200}
-                resp = jsonify(message)
-                resp.status_code = 200
-                return resp
-            else:
-                message = {"status": 500}
-                resp = jsonify(message)
-                resp.status_code = 500
-                return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify(message)
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/openGCode", methods=["POST"])
-def openGCode():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        f = request.form["selectedGCode"]
-        app.data.console_queue.put(f"selectedGcode={f}")
-        tDir = f.split("/")
-        app.data.config.setValue("Computed Settings", "lastSelectedDirectory", tDir[0])
-        home = app.data.config.getHome()
-        app.data.gcodeFile.filename = home + "/.WebControl/gcode/" + f
-        app.data.config.setValue("Maslow Settings", "openFile", tDir[1])
-        returnVal = app.data.gcodeFile.loadUpdateFile()
-        if returnVal:
-            message = {"status": 200}
-            resp = jsonify(message)
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify(message)
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/saveGCode", methods=["POST"])
-def saveGCode():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        print(request.form)
-        f = request.form["fileName"]
-        d = request.form["selectedDirectory"]
-        app.data.console_queue.put(f"selectedGcode={f}")
-        app.data.config.setValue("Computed Settings", "lastSelectedDirectory", d)
-        home = app.data.config.getHome()
-        returnVal = app.data.gcodeFile.saveFile(f, home + "/.WebControl/gcode/" + d)
-        """
-        tDir = f.split("/")
-        app.data.config.setValue("Computed Settings","lastSelectedDirectory",tDir[0])
-        home = app.data.config.getHome()
-        app.data.gcodeFile.filename = home+"/.WebControl/gcode/" + f
-        app.data.config.setValue("Maslow Settings", "openFile", tDir[1])
-        returnVal = app.data.gcodeFile.loadUpdateFile()
-        """
-        if returnVal:
-            app.data.config.setValue("Maslow Settings", "openFile", f)
-            message = {"status": 200}
-            resp = jsonify(message)
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify(message)
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/openBoard", methods=["POST"])
-def openBoard():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        f = request.form["selectedBoard"]
-        app.data.console_queue.put(f"selectedBoard={f}")
-        tDir = f.split("/")
-        app.data.config.setValue(
-            "Computed Settings", "lastSelectedBoardDirectory", tDir[0]
-        )
-        home = app.data.config.getHome()
-        app.data.gcodeFile.filename = home + "/.WebControl/boards/" + f
-        app.data.config.setValue("Maslow Settings", "openBoardFile", tDir[1])
-        returnVal = app.data.boardManager.loadBoard(home + "/.WebControl/boards/" + f)
-        if returnVal:
-            message = {"status": 200}
-            resp = jsonify(message)
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify(message)
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/saveBoard", methods=["POST"])
-def saveBoard():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        print(request.form)
-        f = request.form["fileName"]
-        d = request.form["selectedDirectory"]
-        app.data.console_queue.put(f"selectedBoard={f}")
-        app.data.config.setValue("Computed Settings", "lastSelectedBoardDirectory", d)
-        home = app.data.config.getHome()
-        returnVal = app.data.boardManager.saveBoard(
-            f, home + "/.WebControl/boards/" + d
-        )
-        app.data.config.setValue("Maslow Settings", "openBoardFile", f)
-        if returnVal:
-            message = {"status": 200}
-            resp = jsonify(message)
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify(message)
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/importFile", methods=["POST"])
-def importFile():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        f = request.files["file"]
-        home = app.data.config.getHome()
-        secureFilename = home + "/.WebControl/imports/" + secure_filename(f.filename)
-        f.save(secureFilename)
-        returnVal = app.data.importFile.importGCini(secureFilename)
-        if returnVal:
-            message = {"status": 200}
-            resp = jsonify(message)
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify(message)
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/importFileWCJSON", methods=["POST"])
-def importFileJSON():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        f = request.files["file"]
-        home = app.data.config.getHome()
-        secureFilename = home + "/.WebControl/imports/" + secure_filename(f.filename)
-        f.save(secureFilename)
-        returnVal = app.data.importFile.importWebControlJSON(secureFilename)
-        if returnVal:
-            message = {"status": 200}
-            resp = jsonify(message)
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify(message)
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/importRestoreWebControl", methods=["POST"])
-def importRestoreWebControl():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        f = request.files["file"]
-        home = app.data.config.getHome()
-        secureFilename = home + "/" + secure_filename(f.filename)
-        f.save(secureFilename)
-        returnVal = app.data.actions.restoreWebControl(secureFilename)
-        if returnVal:
-            message = {"status": 200}
-            resp = jsonify(message)
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify(message)
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/sendGCode", methods=["POST"])
-def sendGcode():
-    app.data.logger.resetIdler()
-    # print(request.form)#["gcodeInput"])
-    if request.method == "POST":
-        returnVal = app.data.actions.sendGCode(request.form["gcode"].rstrip())
-        if returnVal:
-            message = {"status": 200}
-            resp = jsonify("success")
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify("failed")
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/triangularCalibration", methods=["POST"])
-def triangularCalibration():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        result = request.form
-        (
-            motorYoffsetEst,
-            rotationRadiusEst,
-            chainSagCorrectionEst,
-            cut34YoffsetEst,
-        ) = app.data.actions.calibrate(result)
-        # print(returnVal)
-        if motorYoffsetEst:
-            message = {
-                "status": 200,
-                "data": {
-                    "motorYoffset": motorYoffsetEst,
-                    "rotationRadius": rotationRadiusEst,
-                    "chainSagCorrection": chainSagCorrectionEst,
-                    "calibrationError": cut34YoffsetEst,
-                },
-            }
-            resp = jsonify(message)
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify(message)
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/holeyCalibration", methods=["POST"])
-def holeyCalibration():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        result = request.form
-        (
-            motorYoffsetEst,
-            distanceBetweenMotors,
-            leftChainTolerance,
-            rightChainTolerance,
-            calibrationError,
-        ) = app.data.actions.holeyCalibrate(result)
-        # print(returnVal)
-        if motorYoffsetEst:
-            message = {
-                "status": 200,
-                "data": {
-                    "motorYoffset": motorYoffsetEst,
-                    "distanceBetweenMotors": distanceBetweenMotors,
-                    "leftChainTolerance": leftChainTolerance,
-                    "rightChainTolerance": rightChainTolerance,
-                    "calibrationError": calibrationError,
-                },
-            }
-            resp = jsonify(message)
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify(message)
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/opticalCalibration", methods=["POST"])
-def opticalCalibration():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        result = request.form
-        message = {"status": 200}
-        resp = jsonify(message)
-        resp.status_code = 200
-        return resp
-    else:
-        message = {"status": 500}
-        resp = jsonify(message)
-        resp.status_code = 500
-        return resp
-
-
-@app.route("/quickConfigure", methods=["POST"])
-def quickConfigure():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        result = request.form
-        app.data.config.updateQuickConfigure(result)
-        message = {"status": 200}
-        resp = jsonify(message)
-        resp.status_code = 200
-        return resp
-
-
-@app.route("/editGCode", methods=["POST"])
-def editGCode():
-    app.data.logger.resetIdler()
-    # print(request.form["gcode"])
-    if request.method == "POST":
-        returnVal = app.data.actions.updateGCode(request.form["gcode"].rstrip())
-        if returnVal:
-            message = {"status": 200}
-            resp = jsonify("success")
-            resp.status_code = 200
-            return resp
-        else:
-            message = {"status": 500}
-            resp = jsonify("failed")
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/downloadDiagnostics", methods=["GET"])
-def downloadDiagnostics():
-    app.data.logger.resetIdler()
-    if request.method == "GET":
-        returnVal = app.data.actions.downloadDiagnostics()
-        if returnVal != False:
-            print(returnVal)
-            return send_file(returnVal, as_attachment=True)
-        else:
-            resp = jsonify("failed")
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/backupWebControl", methods=["GET"])
-def backupWebControl():
-    app.data.logger.resetIdler()
-    if request.method == "GET":
-        returnVal = app.data.actions.backupWebControl()
-        if returnVal != False:
-            print(returnVal)
-            return send_file(returnVal)
-        else:
-            resp = jsonify("failed")
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/editBoard", methods=["POST"])
-def editBoard():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        returnVal = app.data.boardManager.editBoard(request.form)
-        if returnVal:
-            resp = jsonify("success")
-            resp.status_code = 200
-            return resp
-        else:
-            resp = jsonify("failed")
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/trimBoard", methods=["POST"])
-def trimBoard():
-    app.data.logger.resetIdler()
-    if request.method == "POST":
-        returnVal = app.data.boardManager.trimBoard(request.form)
-        if returnVal:
-            resp = jsonify("success")
-            resp.status_code = 200
-            return resp
-        else:
-            resp = jsonify("failed")
-            resp.status_code = 500
-            return resp
-
-
-@app.route("/assets/<path:path>")
-def sendDocs(path):
-    print(path)
-    return send_from_directory("docs/assets/", path)
-
-
-@socketio.on("checkInRequested", namespace="/WebMCP")
-def checkInRequested():
-    socketio.emit("checkIn", namespace="/WebMCP")
-
-
-@socketio.on("connect", namespace="/WebMCP")
-def watchdog_connect():
-    app.data.console_queue.put("watchdog connected")
-    app.data.console_queue.put(request.sid)
-    socketio.emit("connect", namespace="/WebMCP")
-    if app.mcpthread == None:
-        app.data.console_queue.put("going to start mcp thread")
-        app.mcpthread = socketio.start_background_task(
-            app.data.mcpProcessor.start, current_app._get_current_object()
-        )
-        app.data.console_queue.put("created mcp thread")
-        app.mcpthread.start()
-        app.data.console_queue.put("started mcp thread")
-
-
-@socketio.on("my event", namespace="/MaslowCNC")
-def my_event(msg):
-    app.data.console_queue.put(msg["data"])
-
-
-@socketio.on("modalClosed", namespace="/MaslowCNC")
-def modalClosed(msg):
-    app.data.logger.resetIdler()
-    data = json.dumps({"title": msg["data"]})
-    socketio.emit(
-        "message",
-        {"command": "closeModals", "data": data, "dataFormat": "json"},
-        namespace="/MaslowCNC",
-    )
-
-
-@socketio.on("contentModalClosed", namespace="/MaslowCNC")
-def contentModalClosed(msg):
-    # Note, this shouldn't be called anymore
-    # todo: cleanup
-    app.data.logger.resetIdler()
-    data = json.dumps({"title": msg["data"]})
-    print(data)
-    # socketio.emit("message", {"command": "closeContentModals", "data": data, "dataFormat": "json"},
-    #              namespace="/MaslowCNC", )
-
-
-"""
-todo: cleanup
-not used
-@socketio.on("actionModalClosed", namespace="/MaslowCNC")
-def actionModalClosed(msg):
-    app.data.logger.resetIdler()
-    data = json.dumps({"title": msg["data"]})
-    socketio.emit("message", {"command": "closeActionModals", "data": data, "dataFormat": "json"},
-                  namespace="/MaslowCNC", )
-"""
-
-
-@socketio.on("alertModalClosed", namespace="/MaslowCNC")
-def alertModalClosed(msg):
-    app.data.logger.resetIdler()
-    data = json.dumps({"title": msg["data"]})
-    socketio.emit(
-        "message",
-        {"command": "closeAlertModals", "data": data, "dataFormat": "json"},
-        namespace="/MaslowCNC",
-    )
-
-
-@socketio.on("requestPage", namespace="/MaslowCNC")
-def requestPage(msg):
-    print(f"requestPage: {msg}")
-    app.data.logger.resetIdler()
-    app.data.console_queue.put(request.sid)
-    client = request.sid
-    try:
-        (
-            page,
-            title,
-            isStatic,
-            modalSize,
-            modalType,
-            resume,
-        ) = app.webPageProcessor.createWebPage(
-            msg["data"]["page"], msg["data"]["isMobile"], msg["data"]["args"]
-        )
-        # if msg["data"]["page"] != "help":
-        #    client = "all"
-        data = json.dumps(
-            {
-                "title": title,
-                "message": page,
-                "isStatic": isStatic,
-                "modalSize": modalSize,
-                "modalType": modalType,
-                "resume": resume,
-                "client": client,
-            }
-        )
-        socketio.emit(
-            "message",
-            {"command": "activateModal", "data": data, "dataFormat": "json"},
-            namespace="/MaslowCNC",
-        )
-    except Exception as e:
-        app.data.console_queue.put(e)
-
-
-@socketio.on("connect", namespace="/MaslowCNC")
-def test_connect():
-    app.data.console_queue.put("connected")
-    app.data.console_queue.put(request.sid)
-    if app.uithread == None:
-        app.uithread = socketio.start_background_task(
-            app.UIProcessor.start, current_app._get_current_object()
-        )
-        app.uithread.start()
-
-    if not app.data.connectionStatus:
-        app.data.console_queue.put(
-            "Attempting to re-establish connection to controller"
-        )
-        app.data.serialPort.openConnection()
-
-    socketio.emit("my response", {"data": "Connected", "count": 0})
-    address = app.data.hostAddress
-    data = json.dumps({"hostAddress": address})
-    print(data)
-    socketio.emit(
-        "message",
-        {"command": "hostAddress", "data": data, "dataFormat": "json"},
-        namespace="/MaslowCNC",
-    )
-    if app.data.pyInstallUpdateAvailable:
-        app.data.ui_queue1.put("Action", "pyinstallUpdate", "on")
-
-
-@socketio.on("disconnect", namespace="/MaslowCNC")
-def test_disconnect():
-    app.data.console_queue.put("Client disconnected")
-
-
-@socketio.on("action", namespace="/MaslowCNC")
-def command(msg):
-    app.data.logger.resetIdler()
-    retval = app.data.actions.processAction(msg)
-    if retval == "Shutdown":
-        print("Shutting Down")
-        socketio.stop()
-        print("Shutdown")
-    if retval == "TurnOffRPI":
-        print("Turning off RPI")
-        os.system("sudo poweroff")
-
-
-@socketio.on("settingRequest", namespace="/MaslowCNC")
-def settingRequest(msg):
-    app.data.logger.resetIdler()
-    # didn't move to actions.. this request is just to send it computed values.. keeping it here makes it faster than putting it through the UIProcessor
-    setting, value = app.data.actions.processSettingRequest(
-        msg["data"]["section"], msg["data"]["setting"]
-    )
-    if setting is not None:
-        data = json.dumps({"setting": setting, "value": value})
-        socketio.emit(
-            "message",
-            {"command": "requestedSetting", "data": data, "dataFormat": "json"},
-            namespace="/MaslowCNC",
-        )
-
-
-@socketio.on("updateSetting", namespace="/MaslowCNC")
-def updateSetting(msg):
-    app.data.logger.resetIdler()
-    if not app.data.actions.updateSetting(msg["data"]["setting"], msg["data"]["value"]):
-        app.data.ui_queue1.put("Alert", "Alert", "Error updating setting")
-
-
-@socketio.on("checkForGCodeUpdate", namespace="/MaslowCNC")
-def checkForGCodeUpdate(msg):
-    app.data.logger.resetIdler()
-    # this currently doesn't check for updated gcode, it just resends it..
-    ## the gcode file might change the active units so we need to inform the UI of the change.
-    app.data.ui_queue1.put("Action", "unitsUpdate", "")
-    app.data.ui_queue1.put("Action", "gcodeUpdate", "")
-
-
-@socketio.on("checkForBoardUpdate", namespace="/MaslowCNC")
-def checkForBoardUpdate(msg):
-    app.data.logger.resetIdler()
-    # this currently doesn't check for updated board, it just resends it..
-    app.data.ui_queue1.put("Action", "boardUpdate", "")
-
-
-@socketio.on("connect", namespace="/MaslowCNCLogs")
-def log_connect():
-    app.data.console_queue.put("connected to log")
-    app.data.console_queue.put(request.sid)
-    if app.logstreamerthread == None:
-        app.logstreamerthread = socketio.start_background_task(
-            app.LogStreamer.start, current_app._get_current_object()
-        )
-        app.logstreamerthread.start()
-
-    socketio.emit(
-        "my response", {"data": "Connected", "count": 0}, namespace="/MaslowCNCLog"
-    )
-
-
-@socketio.on("disconnect", namespace="/MaslowCNCLogs")
-def log_disconnect():
-    app.data.console_queue.put("Client disconnected")
+init_route(app)
+init_socket_mcp(app)
+init_socket_maslowcnc(app)
+init_socket_maslowcnclogs(app)
 
 
 @app.template_filter("isnumber")
@@ -877,12 +103,13 @@ if __name__ == "__main__":
     app.data.releaseManager.processAbsolutePath(os.path.abspath(__file__))
     print("-$$$$$-")
 
-    print("opening browser")
-    webPortStr = str(webPortInt)
-    webbrowser.open_new_tab("http://localhost:" + webPortStr)
+    webHost = "http://localhost"
+    print(f"opening browser on {webHost}:{webPortInt}")
+    webbrowser.open_new_tab(f"{webHost}:{webPortInt}")
     host_name = socket.gethostname()
     host_ip = socket.gethostbyname(host_name)
-    app.data.hostAddress = host_ip + ":" + webPortStr
+    print(f"setting app data host address to {host_ip}:{webPortInt}")
+    app.data.hostAddress = f"{host_ip}:{webPortInt}"
 
     # app.data.shutdown = shutdown
     socketio.run(app, use_reloader=False, host="0.0.0.0", port=webPortInt)
